[TAN-2343] Move code into core
