+++ back/app/services/geojson_export/domicile_field_for_geojson.rb
@@ -1,0 +1,9 @@
+module GeojsonExport
+  class DomicileFieldForGeojson < CustomFieldForGeojson
+    private
+
+    def option_index
+      @option_index ||= Area.all.index_by(&:id)
+    end
+  end
+end
