--- back/engines/commercial/custom_maps/app/policies/custom_maps/patches/idea_custom_fields/idea_custom_field_policy.rb
@@ -1,11 +1,0 @@
-module CustomMaps
-  module Patches
-    module IdeaCustomFields
-      module IdeaCustomFieldPolicy
-        def as_geojson?
-          update_all?
-        end
-      end
-    end
-  end
-end
