+++ back/engines/commercial/idea_custom_fields/app/controllers/idea_custom_fields/web_api/v1/admin/idea_custom_fields_controller.rb
@@ -23,7 +23,7 @@ module IdeaCustomFields
       }
     }
 
-    before_action :set_custom_field, only: %i[show]
+    before_action :set_custom_field, only: %i[show as_geojson]
     before_action :set_custom_form, only: %i[index update_all]
     skip_after_action :verify_policy_scoped
     rescue_from UpdatingFormWithInputError, with: :render_updating_form_with_input_error
@@ -50,6 +50,18 @@ module IdeaCustomFields
       ).serializable_hash
     end
 
+    def as_geojson
+      # set_custom_field
+      raise_error_if_not_geographic_field
+
+      phase = Phase.find(params[:phase_id])
+      geojson = I18n.with_locale(current_user.locale) do
+        GeojsonExport::GeojsonGenerator.new(phase, @custom_field).generate_geojson
+      end
+
+      send_data geojson, type: 'application/json', filename: geojson_filename(phase)
+    end
+
     def update_all
       authorize CustomField.new(resource: @custom_form), :update_all?, policy_class: IdeaCustomFieldPolicy
       @participation_method = @custom_form.participation_context.pmethod
@@ -74,6 +86,17 @@ module IdeaCustomFields
     # Overriden from CustomMaps::Patches::IdeaCustomFields::WebApi::V1::Admin::IdeaCustomFieldsController
     def include_in_index_response
       %i[options options.image]
+    end
+
+    def raise_error_if_not_geographic_field
+      return if @custom_field.geographic_input?
+
+      raise "Custom field with input_type: '#{@custom_field.input_type}' is not a geographic type"
+    end
+
+    def geojson_filename(phase)
+      "#{MultilocService.new(app_configuration: @app_configuration).t(phase.title_multiloc).tr(' ', '_')}" \
+        "_#{Time.now.strftime('%Y-%m-%d')}.geojson"
     end
 
     def update_fields!(page_temp_ids_to_ids_mapping, option_temp_ids_to_ids_mapping, errors)
