+++ back/engines/commercial/idea_custom_fields/config/routes.rb
@@ -23,6 +23,7 @@ IdeaCustomFields::Engine.routes.draw do
             defaults: { container_type: 'Phase' }
           ) do
             patch 'update_all', on: :collection
+            get :as_geojson, on: :member, action: 'as_geojson'
             resources :custom_field_options, controller: '/web_api/v1/custom_field_options', only: %i[show]
           end
         end
