+++ back/engines/commercial/idea_custom_fields/spec/acceptance/idea_custom_fields/phase_context/as_geojson_spec.rb
