+++ back/spec/services/geojson_export/geojson_value_visitor_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe CustomMaps::GeojsonExport::GeojsonValueVisitor do
+describe GeojsonExport::GeojsonValueVisitor do
   subject(:visitor) do
     described_class.new(model, option_index)
   end
